Parquet: Refactor BaseParquetWriter
